Made the content load per AsyncTask
